#611 Moved Bearer and API Key security scheme tests to OpenApiIntegrationTest class and grouped them in individual inner classes.
