+++ core/src/test/kotlin/in/specmatic/conversions/OpenApiIntegrationTest.kt
@@ -2,11 +2,14 @@ package `in`.specmatic.conversions
 
 import com.google.common.net.HttpHeaders
 import `in`.specmatic.core.*
+import `in`.specmatic.core.pattern.ContractException
 import `in`.specmatic.core.value.Value
+import `in`.specmatic.stub.createStubFromContracts
 import `in`.specmatic.test.TestExecutor
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 import java.io.File
 
 class OpenApiIntegrationTest {
@@ -350,6 +353,260 @@ Given openapi openapi/hello_with_oauth2_resource_owner_password_credentials_flow
                 assertThat(result).isInstanceOf(Result.Success::class.java)
             }
 
+        }
+    }
+
+    @Nested
+    inner class BearerSecuritySchemeTest {
+        @Test
+        fun `should generate test with bearer security scheme with authorization header value from example`() {
+            val contract: Feature = parseGherkinStringToFeature(
+                """
+Feature: Authenticated
+
+  Background:
+    Given openapi openapi/authenticated.yaml
+  
+  Scenario: Bearer auth test
+    When GET /hello/(id:number)
+    Then status 200
+    
+    Examples:
+    | Authorization | id |
+    | Bearer abc123 | 10 |
+        """.trimIndent(), sourceSpecPath
+            )
+
+            val contractTests = contract.generateContractTestScenarios(emptyList())
+            val result = executeTest(contractTests.single(), object : TestExecutor {
+                override fun execute(request: HttpRequest): HttpResponse {
+                    assertThat(request.headers).containsEntry("Authorization", "Bearer abc123")
+                    return HttpResponse.OK("success")
+                }
+
+                override fun setServerState(serverState: Map<String, Value>) {
+
+                }
+
+            })
+
+            assertThat(result).isInstanceOf(Result.Success::class.java)
+        }
+
+        @Test
+        fun `should generate test with bearer security scheme with random token in authorization header when no example exists`() {
+            val feature = parseContractFileToFeature("./src/test/resources/openapi/authenticated.yaml")
+            val contractTests = feature.generateContractTestScenarios(emptyList())
+            var requestMadeWithRandomlyGeneratedBearerToken = false
+            contractTests.forEach { scenario ->
+                val result = executeTest(scenario, object : TestExecutor {
+                    override fun execute(request: HttpRequest): HttpResponse {
+                        request.headers[HttpHeaders.AUTHORIZATION]?.takeIf {
+                            it.matches(Regex("Bearer (\\S+)"))
+                        }?.let {
+                            requestMadeWithRandomlyGeneratedBearerToken = true
+                        }
+                        return HttpResponse.OK("success")
+                    }
+
+                    override fun setServerState(serverState: Map<String, Value>) {
+
+                    }
+                })
+                assertThat(result).isInstanceOf(Result.Success::class.java)
+            }
+            assertThat(requestMadeWithRandomlyGeneratedBearerToken).isTrue
+        }
+
+        @Test
+        fun `should match http request with authorization header for spec with bearer security scheme`() {
+            val feature = parseGherkinStringToFeature(
+                """
+Feature: Hello world
+
+Background:
+  Given openapi openapi/hello.yaml
+        """.trimIndent(), sourceSpecPath
+            )
+            val httpRequest = HttpRequest(
+                "GET",
+                "/hello/1",
+                mapOf(
+                    HttpHeaders.AUTHORIZATION to "Bearer foo"
+                )
+            )
+            val result = feature.scenarios.first().httpRequestPattern.matches(httpRequest, Resolver())
+            assertThat(result).isInstanceOf(Result.Success::class.java)
+        }
+
+        @Test
+        fun `should not match http request without authorization header for spec with bearer security scheme`() {
+            val feature = parseGherkinStringToFeature(
+                """
+Feature: Hello world
+
+Background:
+  Given openapi openapi/hello.yaml
+        """.trimIndent(), sourceSpecPath
+            )
+            val httpRequest = HttpRequest(
+                "GET",
+                "/hello/1"
+            )
+            val result = feature.scenarios.first().httpRequestPattern.matches(httpRequest, Resolver())
+            assertThat(result).isInstanceOf(Result.Failure::class.java)
+            assertThat(result.reportString()).contains("Authorization header is missing in request")
+        }
+
+        @Test
+        fun `should not match http request with authorization header but without the bearer prefix for spec with bearer security scheme`() {
+            val feature = parseGherkinStringToFeature(
+                """
+Feature: Hello world
+
+Background:
+  Given openapi openapi/hello.yaml
+        """.trimIndent(), sourceSpecPath
+            )
+            val httpRequest = HttpRequest(
+                "GET",
+                "/hello/1",
+                mapOf(
+                    HttpHeaders.AUTHORIZATION to "foo"
+                )
+            )
+            val result = feature.scenarios.first().httpRequestPattern.matches(httpRequest, Resolver())
+            assertThat(result).isInstanceOf(Result.Failure::class.java)
+            assertThat(result.reportString()).contains("Authorization header must be prefixed with \"Bearer\"")
+        }
+
+    }
+
+    @Nested
+    inner class ApiKeySecuritySchemeTest {
+        @Test
+        fun `should generate test with query param api key auth security scheme value from row`() {
+            val contract: Feature = parseGherkinStringToFeature(
+                """
+Feature: Authenticated
+
+  Background:
+    Given openapi openapi/authenticated.yaml
+  
+  Scenario: Query param auth test
+    When GET /hello/(id:number)
+    Then status 200
+    
+    Examples:
+    | apiKey | id |
+    | abc123 | 10 |
+        """.trimIndent(), sourceSpecPath
+            )
+
+            val contractTests = contract.generateContractTestScenarios(emptyList())
+            val result = executeTest(contractTests.single(), object : TestExecutor {
+                override fun execute(request: HttpRequest): HttpResponse {
+                    assertThat(request.queryParams).containsEntry("apiKey", "abc123")
+                    return HttpResponse.OK("success")
+                }
+
+                override fun setServerState(serverState: Map<String, Value>) {
+
+                }
+
+            })
+
+            assertThat(result).isInstanceOf(Result.Success::class.java)
+        }
+
+        @Test
+        fun `should throw not supported error when a security scheme other than oauth2 bearer or api key is defined`() {
+            assertThrows<ContractException> {
+                parseGherkinStringToFeature(
+                    """
+Feature: Hello world
+
+Background:
+  Given openapi openapi/unsupported-authentication.yaml
+        """.trimIndent(), sourceSpecPath
+                )
+            }.also { assertThat(it.message).isEqualTo("Specmatic only supports oauth2, bearer, and api key authentication (header, query) security schemes at the moment") }
+        }
+
+        @Test
+        fun `should throw not supported error when a non-query-or-header API security scheme is defined`() {
+            assertThrows<ContractException> {
+                parseGherkinStringToFeature(
+                    """
+Feature: Hello world
+
+Background:
+  Given openapi openapi/apiKeyAuthCookie.yaml
+        """.trimIndent(), sourceSpecPath
+                )
+            }.also { assertThat(it.message).isEqualTo("Specmatic only supports oauth2, bearer, and api key authentication (header, query) security schemes at the moment") }
+        }
+
+        @Test
+        fun `should generate test with api key security scheme value from row`() {
+            val contract: Feature = parseGherkinStringToFeature(
+                """
+Feature: Authenticated
+
+  Background:
+    Given openapi openapi/authenticated.yaml
+  
+  Scenario: Header auth test
+    When GET /hello/(id:number)
+    Then status 200
+    
+    Examples:
+    | X-API-KEY | id |
+    | abc123    | 10 |
+        """.trimIndent(), sourceSpecPath
+            )
+
+            val contractTests = contract.generateContractTestScenarios(emptyList())
+            val result = executeTest(contractTests.single(), object : TestExecutor {
+                override fun execute(request: HttpRequest): HttpResponse {
+                    assertThat(request.headers).containsEntry("X-API-KEY", "abc123")
+                    return HttpResponse.OK("success")
+                }
+
+                override fun setServerState(serverState: Map<String, Value>) {
+
+                }
+
+            })
+
+            assertThat(result).isInstanceOf(Result.Success::class.java)
+        }
+
+        @Test
+        fun `should generate stub that authenticates with api key in header and query`() {
+            createStubFromContracts(listOf("src/test/resources/openapi/apiKeyAuth.yaml")).use {
+                val requestWithHeader = HttpRequest(
+                    method = "GET",
+                    path = "/hello/10",
+                    headers = mapOf(
+                        "X-API-KEY" to "test"
+                    )
+                )
+
+                val responseFromHeader = it.client.execute(requestWithHeader)
+                assertThat(responseFromHeader.status).isEqualTo(200)
+
+                val requestWithQuery = HttpRequest(
+                    method = "GET",
+                    path = "/hello/10",
+                    queryParams = mapOf(
+                        "apiKey" to "test"
+                    )
+                )
+
+                val responseFromQuery = it.client.execute(requestWithQuery)
+                assertThat(responseFromQuery.status).isEqualTo(200)
+            }
         }
     }
