+++ core/src/test/kotlin/in/specmatic/conversions/OpenApiKtTest.kt
@@ -778,96 +778,6 @@ Background:
         }
     }
 
-    @Test
-    fun `should generate stub that authenticates with api key in header and query`() {
-        createStubFromContracts(listOf("src/test/resources/openapi/apiKeyAuth.yaml")).use {
-            val requestWithHeader = HttpRequest(
-                method = "GET",
-                path = "/hello/10",
-                headers = mapOf(
-                    "X-API-KEY" to "test"
-                )
-            )
-
-            val responseFromHeader = it.client.execute(requestWithHeader)
-            assertThat(responseFromHeader.status).isEqualTo(200)
-
-            val requestWithQuery = HttpRequest(
-                method = "GET",
-                path = "/hello/10",
-                queryParams = mapOf(
-                    "apiKey" to "test"
-                )
-            )
-
-            val responseFromQuery = it.client.execute(requestWithQuery)
-            assertThat(responseFromQuery.status).isEqualTo(200)
-        }
-    }
-
-    @Test
-    fun `should throw not supported error when a security scheme other than oauth2 bearer or api key is defined`() {
-        assertThrows<ContractException> {
-            parseGherkinStringToFeature(
-                """
-Feature: Hello world
-
-Background:
-  Given openapi openapi/unsupported-authentication.yaml
-        """.trimIndent(), sourceSpecPath
-            )
-        }.also { assertThat(it.message).isEqualTo("Specmatic only supports oauth2, bearer, and api key authentication (header, query) security schemes at the moment") }
-    }
-
-    @Test
-    fun `should throw not supported error when a non-query-or-header API security scheme is defined`() {
-        assertThrows<ContractException> {
-            parseGherkinStringToFeature(
-                """
-Feature: Hello world
-
-Background:
-  Given openapi openapi/apiKeyAuthCookie.yaml
-        """.trimIndent(), sourceSpecPath
-            )
-        }.also { assertThat(it.message).isEqualTo("Specmatic only supports oauth2, bearer, and api key authentication (header, query) security schemes at the moment") }
-    }
-
-    @Test
-    fun `should generate test with api key security scheme value from row`() {
-        val contract: Feature = parseGherkinStringToFeature(
-            """
-Feature: Authenticated
-
-  Background:
-    Given openapi openapi/authenticated.yaml
-  
-  Scenario: Header auth test
-    When GET /hello/(id:number)
-    Then status 200
-    
-    Examples:
-    | X-API-KEY | id |
-    | abc123    | 10 |
-        """.trimIndent(), sourceSpecPath
-        )
-
-        val contractTests = contract.generateContractTestScenarios(emptyList())
-        val result = executeTest(contractTests.single(), object : TestExecutor {
-            override fun execute(request: HttpRequest): HttpResponse {
-                assertThat(request.headers).containsEntry("X-API-KEY", "abc123")
-                return HttpResponse.OK("success")
-            }
-
-            override fun setServerState(serverState: Map<String, Value>) {
-
-            }
-
-        })
-
-        assertThat(result).isInstanceOf(Result.Success::class.java)
-    }
-
     @ParameterizedTest
     @MethodSource("multiPartFileUploadSpecs")
     fun `should generate test with multipart file upload`(openApiFile: String, fileName: String) {
@@ -959,138 +869,6 @@ Background:
             }
             assertThat(httpClientErrorException.message).contains("The contract expected a file, but got content instead.")
         }
-    }
-
-    @Test
-    fun `should generate test with bearer auth security scheme value from row`() {
-        val contract: Feature = parseGherkinStringToFeature(
-            """
-Feature: Authenticated
-
-  Background:
-    Given openapi openapi/authenticated.yaml
-  
-  Scenario: Bearer auth test
-    When GET /hello/(id:number)
-    Then status 200
-    
-    Examples:
-    | Authorization | id |
-    | Bearer abc123 | 10 |
-        """.trimIndent(), sourceSpecPath
-        )
-
-        val contractTests = contract.generateContractTestScenarios(emptyList())
-        val result = executeTest(contractTests.single(), object : TestExecutor {
-            override fun execute(request: HttpRequest): HttpResponse {
-                assertThat(request.headers).containsEntry("Authorization", "Bearer abc123")
-                return HttpResponse.OK("success")
-            }
-
-            override fun setServerState(serverState: Map<String, Value>) {
-
-            }
-
-        })
-
-        assertThat(result).isInstanceOf(Result.Success::class.java)
-    }
-
-    @Test
-    fun `should match http request with authorization header for spec with bearer security scheme`() {
-        val feature = parseGherkinStringToFeature(
-            """
-Feature: Hello world
-
-Background:
-  Given openapi openapi/hello.yaml
-        """.trimIndent(), sourceSpecPath
-        )
-        val httpRequest = HttpRequest(
-            "GET",
-            "/hello/1",
-            mapOf(
-                AUTHORIZATION to "Bearer foo"
-            )
-        )
-        val result = feature.scenarios.first().httpRequestPattern.matches(httpRequest, Resolver())
-        assertThat(result).isInstanceOf(Result.Success::class.java)
-    }
-
-    @Test
-    fun `should not match http request without authorization header for spec with bearer security scheme`() {
-        val feature = parseGherkinStringToFeature(
-            """
-Feature: Hello world
-
-Background:
-  Given openapi openapi/hello.yaml
-        """.trimIndent(), sourceSpecPath
-        )
-        val httpRequest = HttpRequest(
-            "GET",
-            "/hello/1"
-        )
-        val result = feature.scenarios.first().httpRequestPattern.matches(httpRequest, Resolver())
-        assertThat(result).isInstanceOf(Result.Failure::class.java)
-        assertThat(result.reportString()).contains("Authorization header is missing in request")
-    }
-
-    @Test
-    fun `should not match http request with authorization header but without the bearer prefix for spec with bearer security scheme`() {
-        val feature = parseGherkinStringToFeature(
-            """
-Feature: Hello world
-
-Background:
-  Given openapi openapi/hello.yaml
-        """.trimIndent(), sourceSpecPath
-        )
-        val httpRequest = HttpRequest(
-            "GET",
-            "/hello/1",
-            mapOf(
-                AUTHORIZATION to "foo"
-            )
-        )
-        val result = feature.scenarios.first().httpRequestPattern.matches(httpRequest, Resolver())
-        assertThat(result).isInstanceOf(Result.Failure::class.java)
-        assertThat(result.reportString()).contains("Authorization header must be prefixed with \"Bearer\"")
-    }
-
-    @Test
-    fun `should generate test with query param api key auth security scheme value from row`() {
-        val contract: Feature = parseGherkinStringToFeature(
-            """
-Feature: Authenticated
-
-  Background:
-    Given openapi openapi/authenticated.yaml
-  
-  Scenario: Query param auth test
-    When GET /hello/(id:number)
-    Then status 200
-    
-    Examples:
-    | apiKey | id |
-    | abc123 | 10 |
-        """.trimIndent(), sourceSpecPath
-        )
-
-        val contractTests = contract.generateContractTestScenarios(emptyList())
-        val result = executeTest(contractTests.single(), object : TestExecutor {
-            override fun execute(request: HttpRequest): HttpResponse {
-                assertThat(request.queryParams).containsEntry("apiKey", "abc123")
-                return HttpResponse.OK("success")
-            }
-
-            override fun setServerState(serverState: Map<String, Value>) {
-
-            }
-
-        })
-
-        assertThat(result).isInstanceOf(Result.Success::class.java)
     }
 
     @Test
